server/requirements.txt
Flask==1.1.2
mysql-connector-python
# mysql-connector==2.2.9

